chore(nextjs): Remove cyclic module dependency
